Used int64 instead of int. Fixed some minor issues and renamed Multiprocess to Threading

+++ api/types/int64/Int64.sol
@@ -2,7 +2,7 @@ pragma solidity ^0.5.0;
 
 import "./Base.sol";
 
-contract Integer {
+contract Int64 {
     Base base;
 
     constructor  () public {  base = new Base(); }
@@ -12,15 +12,15 @@ contract Integer {
         return abi.decode(abi.decode(base.pop(), (bytes)), (int));  
     }
 
-    function push(int elem) public { //9e c6 69 25
+    function push(int64 elem) public { //9e c6 69 25
        base.push(abi.encodeWithSignature("push(bytes,bytes)",  base.id(), abi.encode(elem)));
     }   
 
-    function get(uint256 idx) public returns(int)  { // 31 fe 88 d0
-        return abi.decode(abi.decode(base.get(idx), (bytes)), (int));
+    function get(uint256 idx) public returns(int64)  { // 31 fe 88 d0
+        return abi.decode(abi.decode(base.get(idx), (bytes)), (int64));
     }
 
-    function set(uint256 idx, int elem) public { // 7a fa 62 38
+    function set(uint256 idx, int64 elem) public { // 7a fa 62 38
         base.set(abi.encodeWithSignature("set(bytes,uint256,bytes)", base.id(), idx, abi.encode(elem)));        
     }
 }
